gateway/requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
boto3==1.28.64
botocore==1.31.64
build==1.0.3
certifi==2023.7.22
click==8.1.7
distlib==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
packaging==23.2
pip-tools==7.3.0
pipenv==2023.10.3
platformdirs==3.11.0
protobuf==4.24.4
pydantic==2.4.2
pydantic_core==2.10.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.5
watchfiles==0.20.0
websockets==11.0.3
requests~=2.31.0